+++ src/string/index.ts
@@ -1,2 +1,3 @@
 export * from './is-aws-arn'
 export * from './is-aws-region'
+export * from './not-matches'