+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-   `NotMatches`: Added new validator.
+
 ## [3.0.0] - 2023-03-28
 
 ### Removed